1.Creating Matrix and viewing it in Lexical Scoping Process:  This progrm calculates inverse of a special matrix that is created by calling set() function of the makeCacheMatrix(). Once the matrix is created you may view it by calling get() function.
2. Calculating Inverse:
Once the matrix is created it can be put as arguments while calling cacheSolve() function that calculates the inverse using the solve() function.

This commit shows two changes where I put comments to show the location of the command that calculates inverse of the matrix.
Another purpose of this commit is to give a detail explanation of how the program works.

To run the program (command example)

invfun <- makeCacheMatrix()
invfun$set(matrix(i:j), k)  ## Must be an invertible matrix
cacheSolve(invfun)
invfun$getinv